+++ CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.2.0](https://github.com/carnojs/carno.js/compare/v1.1.2...v1.2.0) (2026-02-26)
+
+
+### Features
+
+* Add CompressionMiddleware for response compression and enhance middleware handling ([eba0f96](https://github.com/carnojs/carno.js/commit/eba0f96a0143eff7baf53d61c138fa8947f2865b))
+* Implement ManyToMany and OneToOne relationship decorators ([f0eae30](https://github.com/carnojs/carno.js/commit/f0eae30a49f2ec354a8f9bcac62a582f1f0077f0))
+
+
+
+
+
 ## [1.0.8](https://github.com/carnojs/carno.js/compare/v1.0.7...v1.0.8) (2026-01-13)
 
 **Note:** Version bump only for package @carno.js/core